+++ src/Runner/Assert.php
@@ -131,7 +131,7 @@ final class Assert
     }
 
     /**
-     * @param 0|positive-int $expected
+     * @param int<0, max> $expected
      * @param non-empty-string $message
      *
      * @throws Failure