+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 6.11.0 - 2026-03-18
+
+### Added
+
+- `Innmind\BlackBox\Application::useExhaustiveShrinking()`
+
+### Changed
+
+- By default now the shrinking stops when the type of error changes
+
 ## 6.10.0 - 2026-02-10
 
 ### Added